backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend of travel app"
authors = [
    {name = "Dena"}
]
license = {text = "common"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "amadeus (>=12.0.0,<13.0.0)",
    "google-adk (>=0.3.0,<0.4.0)",
    "litellm (>=1.67.4.post1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
